Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    int x,y;
+    float x,y;
     char z;
-    scanf("%d %d %c", &x,&y,&z);
+    scanf("%f %f %c", &x,&y,&z);
     if(z=='+') {
         printf("%d", x+y);
     } else if(z=='-') {
